fixed routing issue
