+++ pages/signin.tsx
@@ -1,0 +1,104 @@
+import GoogleIcon from '@mui/icons-material/Google';
+import { useState } from 'react';
+import Image from 'next/image';
+import Head from 'next/head';
+import { NextPage } from 'next';
+import { supabase } from '../utils/supabaseClient';
+import { useRouter } from 'next/router';
+
+const SignIn: NextPage = () => {
+	const router = useRouter();
+	const [email, setEmail] = useState('');
+	const [password, setPassword] = useState('');
+	const handleSignIn = async () => {
+		const { user, error } = await supabase.auth.signIn({ email, password });
+		if (error) alert(error.message);
+		else {
+			alert(`Success Login in ${user?.email}`);
+			setEmail('');
+			setPassword('');
+		}
+	};
+	return (
+		<>
+			<Head>
+				<title>Defaang / Welcome</title>
+			</Head>
+			<main className='flex h-screen w-screen'>
+				{/* Left section - main login  */}
+
+				<div className='flex h-screen  w-[100%] flex-col items-center justify-center bg-gray-100 lg:w-[50%]'>
+					{/* Now you can decide different login ways - primarily I am giving google and twiiter for more add it yourself. */}
+					<div className='w-full  bg-white p-4 md:shadow-lg lg:h-[500px] lg:w-[400px]'>
+						<h1 className='p-2 text-3xl font-semibold'>Login ⚡</h1>
+						<p className='p-2 py-4 text-lg font-medium'>See your growth and get consulting support.</p>
+						<div className='flex w-max cursor-pointer items-center space-x-2 rounded-md bg-blue-600 px-5 py-2 text-white hover:bg-blue-500'>
+							<GoogleIcon className='h-5 w-5' color='inherit' />
+
+							<button>Sign in with Google</button>
+						</div>
+						<form
+							className='mt-4'
+							onSubmit={e => {
+								e.preventDefault();
+								handleSignIn();
+							}}
+						>
+							<div className='flex w-[90%] flex-col space-y-2'>
+								<p>Email</p>
+								<input
+									type='text'
+									placeholder='dojo@gmaill.com'
+									required
+									className='input-field outline-none'
+									value={email}
+									onChange={e => setEmail(e.target.value)}
+								/>
+							</div>
+							<div className='flex w-[90%] flex-col space-y-2 '>
+								<p>Password</p>
+								<input
+									type='password'
+									required
+									className='input-field outline-none'
+									placeholder='*****'
+									value={password}
+									onChange={e => setPassword(e.target.value)}
+								/>
+							</div>
+						</form>
+						<button
+							type='submit'
+							className='my-4 flex  w-[90%] cursor-pointer items-center  justify-center rounded-md bg-blue-600 py-2 font-semibold text-white hover:bg-blue-500'
+							onClick={handleSignIn}
+						>
+							Login
+						</button>
+						<p className='py-2 font-semibold'>
+							Not registered yet ?{' '}
+							<span
+								className='cursor-pointer font-bold text-blue-600 hover:underline'
+								onClick={() => router.push('/signup')}
+							>
+								Register here
+							</span>
+						</p>
+					</div>
+				</div>
+				{/* Right Section - Image and just for UI. */}
+
+				<div className='relative hidden w-[50%] lg:inline-flex'>
+					<div className='absolute top-0 left-0 h-[100%] w-[100%] bg-black opacity-[0.5]'></div>
+					<Image
+						src='https://images.unsplash.com/photo-1608306448197-e83633f1261c?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=687&q=80'
+						alt='right-side-image'
+						layout='fill'
+						className='object-cover'
+					/>
+				</div>
+			</main>
+		</>
+	);
+};
+
+export default SignIn;
